requirements.txt
beautifulsoup4==4.12.3
lxml==5.1.0
numpy==1.26.4
openai==1.12.0
pydantic==2.6.1
python-dotenv==1.0.0
regex==2023.6.3
requests==2.31.0
semantic-router==0.0.22
sentencepiece==0.1.99
transformers==4.30.2
typing_extensions==4.9.0